ErrorValues -> error_code, Architecture -> architecture
